# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Виконання тестового завдання в EVO 2022",
        "title": "CSV-APP API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4444",
    "basePath": "/api/v1",
    "paths": {
        "/csv-file": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "show slice transactions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Search by transaction_id",
                        "name": "transaction_id",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "description": "Search by terminal_id (possible to specify several ids at the same time)",
                        "name": "terminal_id",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "accepted",
                            "declined"
                        ],
                        "type": "string",
                        "description": "Search by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "card",
                            "cash"
                        ],
                        "type": "string",
                        "description": "Search by payment_type",
                        "name": "payment_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "date",
                        "example": "2022-08-12",
                        "description": "From date inclusive",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "date",
                        "example": "2022-09-01",
                        "description": "To date not inclusive",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search on the partially specified payment_narrative",
                        "name": "payment_narrative",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show slice transactions",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schemes.Transaction"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemes.Response"
                        }
                    }
                }
            }
        },
        "/csv-file/download": {
            "get": {
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "download csv file to database",
                "responses": {
                    "200": {
                        "description": "return csv file",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemes.Response"
                        }
                    }
                }
            }
        },
        "/csv-file/upload": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "upload csv file to database",
                "parameters": [
                    {
                        "type": "file",
                        "description": "file to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File is uploaded",
                        "schema": {
                            "$ref": "#/definitions/schemes.Response"
                        }
                    },
                    "400": {
                        "description": "transaction already exist",
                        "schema": {
                            "$ref": "#/definitions/schemes.Response"
                        }
                    },
                    "422": {
                        "description": "invalid content type provided",
                        "schema": {
                            "$ref": "#/definitions/schemes.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemes.Response"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Healthcheck"
                ],
                "summary": "show pong",
                "responses": {
                    "200": {
                        "description": "Server is alive",
                        "schema": {
                            "$ref": "#/definitions/schemes.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "schemes.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "schemes.Transaction": {
            "type": "object",
            "properties": {
                "amount_original": {
                    "type": "number"
                },
                "amount_total": {
                    "type": "number"
                },
                "commission_client": {
                    "type": "number"
                },
                "commission_provider": {
                    "type": "number"
                },
                "commission_ps": {
                    "type": "number"
                },
                "date_input": {
                    "type": "string"
                },
                "date_post": {
                    "type": "string"
                },
                "partner_object_id": {
                    "type": "integer"
                },
                "payee_bank_account": {
                    "type": "string"
                },
                "payee_bank_mfo": {
                    "type": "integer"
                },
                "payee_id": {
                    "type": "integer"
                },
                "payee_name": {
                    "type": "string"
                },
                "payment_narrative": {
                    "type": "string"
                },
                "payment_number": {
                    "type": "string"
                },
                "payment_type": {
                    "type": "string"
                },
                "request_id": {
                    "type": "integer"
                },
                "service": {
                    "type": "string"
                },
                "service_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "terminal_id": {
                    "type": "integer"
                },
                "transaction_id": {
                    "type": "integer"
                }
            }
        }
    }
}